backend-flask/requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
pymongo==4.6.1
Werkzeug==3.0.1
